backend/requirements.txt
dj-database-url==3.0.0
Django==5.2.3
django-cors-headers==4.7.0
django-filter==25.1
django-rest-swagger==2.2.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
gunicorn==23.0.0
pillow==11.2.1
psycopg2-binary==2.9.10
python-decouple==3.8
requests==2.32.4
whitenoise==6.9.0
